requirements.txt
bedrock-agentcore
strands-agents
mcp
requests
python-dotenv
boto3